Add `FromGlueRow` derive macro implementation
Introduce the `FromGlueRow` macro to enable conversion of GlueSQL rows into structs. Validate input derives for named fields and handle type mismatches. Support optional, nested, and collection types with recursive resolution.
